Finalized the quest setup
Thanks to ShinyDragonHunter, psf and Jaizu for the debugging help

+++ include/quests.h
@@ -27,7 +27,7 @@
 
 struct SubQuest
 {
-    const u16 id;
+    const u8 id;
 	const u8 *name;
 	const u8 *desc;
 	const u8 *map;
